+++ src/main/java/cech12/extendedmushrooms/block/InfestedFlowerBlock.java
@@ -1,5 +1,6 @@
 package cech12.extendedmushrooms.block;
 
+import net.minecraft.core.Direction;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.Blocks;
@@ -39,4 +40,15 @@ public class InfestedFlowerBlock extends FlowerBlock {
                     blockPos.getZ() + offset.z + (random.nextDouble() * 0.375), 0.0D, 0.0D, 0.0D);
         }
     }
+
+    @Override
+    public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction face) {
+        return 100;
+    }
+
+    @Override
+    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction face) {
+        return 60;
+    }
+
 }