+++ src/main/java/cech12/extendedmushrooms/block/VerticalPlanksBlock.java
@@ -1,25 +1,11 @@
 package cech12.extendedmushrooms.block;
 
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.core.Direction;
-import net.minecraft.core.BlockPos;
-import net.minecraft.world.level.BlockGetter;
 
 public class VerticalPlanksBlock extends Block {
 
     public VerticalPlanksBlock(Properties properties) {
         super(properties);
-    }
-
-    @Override
-    public int getFlammability(BlockState state, BlockGetter world, BlockPos pos, Direction face) {
-        return 20;
-    }
-
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter world, BlockPos pos, Direction face) {
-        return 5;
     }
 
 }