+++ src/main/java/cech12/extendedmushrooms/block/VerticalSlabBlock.java
@@ -165,14 +165,4 @@ public class VerticalSlabBlock extends Block implements SimpleWaterloggedBlock {
 
     }
 
-    @Override
-    public int getFlammability(BlockState state, BlockGetter world, BlockPos pos, Direction face) {
-        return 20;
-    }
-
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter world, BlockPos pos, Direction face) {
-        return 5;
-    }
-
 }